+++ guidata/tests/test_loadsave_hdf5.py
@@ -15,7 +15,7 @@ file.
 
 import os
 
-from guidata.dataset.hdf5io import HDF5Reader, HDF5Writer
+from guidata.dataset.io import HDF5Reader, HDF5Writer
 from guidata.env import execenv
 from guidata.qthelpers import qt_app_context
 from guidata.tests.test_all_items import Parameters