+++ guidata/tests/test_loadsave_json.py
@@ -14,7 +14,7 @@ This script shows how to save in and then reload data from a JSON file.
 
 import os
 
-from guidata.dataset.jsonio import JSONReader, JSONWriter
+from guidata.dataset.io import JSONReader, JSONWriter
 from guidata.env import execenv
 from guidata.qthelpers import qt_app_context
 from guidata.tests.test_all_items import Parameters